Make layouts SSR
