+++ ui/apps/dashboard/src/app/(organization-active)/(dashboard)/settings/integrations/neon/(steps)/layout.tsx
@@ -1,4 +1,4 @@
-import { StepsProvider } from './Context';
+import { StepsProvider } from '@/components/PostgresIntegration/Context';
 
 export default function Layout({ children }: React.PropsWithChildren) {
   return <StepsProvider>{children}</StepsProvider>;
